+++ docs/src/examples/components/Segment/Variations/SegmentExampleInverted.tsx
@@ -3,10 +3,10 @@ import { Segment } from '@stardust-ui/react'
 
 const SegmentExampleInvertedShorthand = () => (
   <div>
-    <Segment color="purple">Colored segment.</Segment>
+    <Segment color="red">Colored segment.</Segment>
     <br />
-    <Segment inverted color="purple">
-      Colored inverted segment
+    <Segment inverted color="red">
+      Colored inverted segment.
     </Segment>
   </div>
 )